Revert "Build vespa zookeeper server variants with zookeeper 3.5.6 and 3.5.8"
