+++ zookeeper-server/CMakeLists.txt
@@ -1,4 +1,3 @@
 # Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 add_subdirectory(zookeeper-server-common)
-add_subdirectory(zookeeper-server-3.5.6)
-add_subdirectory(zookeeper-server-3.5.8)
+add_subdirectory(zookeeper-server-3.5)
