+++ zookeeper-server/pom.xml
@@ -13,8 +13,7 @@
   <version>7-SNAPSHOT</version>
   <modules>
     <module>zookeeper-server-common</module>
-    <module>zookeeper-server-3.5.6</module>
-    <module>zookeeper-server-3.5.8</module>
+    <module>zookeeper-server-3.5</module>
   </modules>
   <dependencies>
     <dependency>
