--- zookeeper-server/zookeeper-server-3.5.8/CMakeLists.txt
@@ -1,4 +1,0 @@
-# Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-install_fat_java_artifact(zookeeper-server-3.5.8)
-# TODO: Needs to be included when this is the wanted default version (and other symlinks need to be removed)
-#install_symlink(lib/jars/zookeeper-server-3.5.8-jar-with-dependencies.jar lib/jars/zookeeper-server-jar-with-dependencies.jar)
