+++ zookeeper-server/zookeeper-server-3.5/pom.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0"?>
-<!-- Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root. -->
+<!-- Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root. -->
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
@@ -8,7 +8,7 @@
     <version>7-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
-  <artifactId>zookeeper-server-3.5.8</artifactId>
+  <artifactId>zookeeper-server-3.5</artifactId>
   <packaging>container-plugin</packaging>
   <version>7-SNAPSHOT</version>
   <dependencies>
@@ -20,7 +20,7 @@
     <dependency>
       <groupId>org.apache.zookeeper</groupId>
       <artifactId>zookeeper</artifactId>
-      <version>3.5.8</version>
+      <version>${zookeeper.server.version}</version>
     </dependency>
     <dependency>
       <groupId>org.slf4j</groupId>
