+++ zookeeper-server/zookeeper-server-3.5/src/main/java/com/yahoo/vespa/zookeeper/VespaZooKeeperServerImpl.java
@@ -1,8 +1,10 @@
-// Copyright Verizon Media. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-
+// Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 package com.yahoo.vespa.zookeeper;
 
+import com.google.inject.Inject;
 import com.yahoo.cloud.config.ZookeeperServerConfig;
+import com.yahoo.component.AbstractComponent;
+import java.util.logging.Level;
 import com.yahoo.security.KeyStoreBuilder;
 import com.yahoo.security.KeyStoreType;
 import com.yahoo.security.KeyStoreUtils;
@@ -11,6 +13,7 @@ import com.yahoo.security.SslContextBuilder;
 import com.yahoo.security.X509CertificateUtils;
 import com.yahoo.security.tls.TlsContext;
 import com.yahoo.security.tls.TransportSecurityOptions;
+import com.yahoo.security.tls.TransportSecurityUtils;
 import com.yahoo.text.Utf8;
 
 import javax.net.ssl.SSLContext;
@@ -30,34 +33,54 @@ import java.util.stream.Collectors;
 
 import static com.yahoo.vespa.defaults.Defaults.getDefaults;
 
-public class Configurator {
+/**
+ * Writes zookeeper config and starts zookeeper server.
+ *
+ * @author Ulf Lilleengen
+ * @author Harald Musum
+ */
+public class VespaZooKeeperServerImpl extends AbstractComponent implements Runnable, VespaZooKeeperServer {
+
+    private static final java.util.logging.Logger log = java.util.logging.Logger.getLogger(VespaZooKeeperServerImpl.class.getName());
     private static final String ZOOKEEPER_JMX_LOG4J_DISABLE = "zookeeper.jmx.log4j.disable";
     static final String ZOOKEEPER_JUTE_MAX_BUFFER = "jute.maxbuffer";
-
+    private final Thread zkServerThread;
     private final ZookeeperServerConfig zookeeperServerConfig;
     private final String configFilePath;
     private final String jksKeyStoreFilePath;
 
-    public Configurator(ZookeeperServerConfig zookeeperServerConfig) {
+    VespaZooKeeperServerImpl(ZookeeperServerConfig zookeeperServerConfig, boolean startServer,
+                            Optional<TransportSecurityOptions> transportSecurityOptions) {
         this.zookeeperServerConfig = zookeeperServerConfig;
-        this.configFilePath = zookeeperServerConfig.zooKeeperConfigFile();
-        this.jksKeyStoreFilePath = zookeeperServerConfig.jksKeyStoreFile();
-        System.setProperty(ZOOKEEPER_JMX_LOG4J_DISABLE, "true");
+        System.setProperty("zookeeper.jmx.log4j.disable", "true");
         System.setProperty("zookeeper.snapshot.trust.empty", Boolean.valueOf(zookeeperServerConfig.trustEmptySnapshot()).toString());
         System.setProperty(ZOOKEEPER_JUTE_MAX_BUFFER, Integer.valueOf(zookeeperServerConfig.juteMaxBuffer()).toString());
+
+        configFilePath = getDefaults().underVespaHome(zookeeperServerConfig.zooKeeperConfigFile());
+        jksKeyStoreFilePath = getDefaults().underVespaHome(zookeeperServerConfig.jksKeyStoreFile());
+        writeConfigToDisk(zookeeperServerConfig, transportSecurityOptions);
+        zkServerThread = new Thread(this, "zookeeper server");
+        if (startServer) {
+            zkServerThread.start();
+        }
     }
 
-    void writeConfigToDisk(Optional<TransportSecurityOptions> transportSecurityOptions) {
+    @Inject
+    public VespaZooKeeperServerImpl(ZookeeperServerConfig zookeeperServerConfig) {
+        this(zookeeperServerConfig, true, TransportSecurityUtils.getOptions());
+    }
+
+    private void writeConfigToDisk(ZookeeperServerConfig config, Optional<TransportSecurityOptions> transportSecurityOptions) {
         new File(configFilePath).getParentFile().mkdirs();
 
         try {
             writeZooKeeperConfigFile(zookeeperServerConfig, transportSecurityOptions);
-            writeMyIdFile(zookeeperServerConfig);
+            writeMyIdFile(config);
             transportSecurityOptions.ifPresent(this::writeJksKeystore);
         } catch (IOException e) {
             throw new RuntimeException("Error writing zookeeper config", e);
         }
-    }
+   }
 
     private void writeZooKeeperConfigFile(ZookeeperServerConfig config,
                                           Optional<TransportSecurityOptions> transportSecurityOptions) throws IOException {
@@ -138,10 +161,31 @@ public class Configurator {
         sb.append("server.").append(server.id()).append("=").append(server.hostname()).append(":").append(server.quorumPort()).append(":").append(server.electionPort()).append("\n");
     }
 
+    private void shutdown() {
+        zkServerThread.interrupt();
+        try {
+            zkServerThread.join();
+        } catch (InterruptedException e) {
+            log.log(Level.WARNING, "Error joining server thread on shutdown", e);
+        }
+    }
 
+    @Override
+    public void run() {
+        System.setProperty(ZOOKEEPER_JMX_LOG4J_DISABLE, "true");
+        String[] args = new String[]{getDefaults().underVespaHome(zookeeperServerConfig.zooKeeperConfigFile())};
+        log.log(Level.INFO, "Starting ZooKeeper server with config file " + args[0] +
+                ". Trying to establish ZooKeeper quorum (members: " + zookeeperServerHostnames(zookeeperServerConfig) + ")");
+        org.apache.zookeeper.server.quorum.QuorumPeerMain.main(args);
+    }
 
+    @Override
+    public void deconstruct() {
+        shutdown();
+        super.deconstruct();
+    }
 
-    static Set<String> zookeeperServerHostnames(ZookeeperServerConfig zookeeperServerConfig) {
+    private static Set<String> zookeeperServerHostnames(ZookeeperServerConfig zookeeperServerConfig) {
         return zookeeperServerConfig.server().stream().map(ZookeeperServerConfig.Server::hostname).collect(Collectors.toSet());
     }
